Move Ractor to an extension library for now
Because so many libraries change behavior upon detecting the
"Ractor" constant, a partial implementation prevents most of them
from booting properly. This moves the definition of Ractor into an
extension library, so it can be left undefined for applications
that will not use Ractors.
