+++ core/src/main/java/org/jruby/ext/ractor/RactorLibrary.java
@@ -1,0 +1,48 @@
+/***** BEGIN LICENSE BLOCK *****
+ * Version: EPL 2.0/GPL 2.0/LGPL 2.1
+ *
+ * The contents of this file are subject to the Eclipse Public
+ * License Version 2.0 (the "License"); you may not use this file
+ * except in compliance with the License. You may obtain a copy of
+ * the License at http://www.eclipse.org/legal/epl-v20.html
+ *
+ * Software distributed under the License is distributed on an "AS
+ * IS" basis, WITHOUT WARRANTY OF ANY KIND, either express or
+ * implied. See the License for the specific language governing
+ * rights and limitations under the License.
+ *
+ * Copyright (C) 2025 Charles Oliver Nutter
+ *
+ * Alternatively, the contents of this file may be used under the terms of
+ * either of the GNU General Public License Version 2 or later (the "GPL"),
+ * or the GNU Lesser General Public License Version 2.1 or later (the "LGPL"),
+ * in which case the provisions of the GPL or the LGPL are applicable instead
+ * of those above. If you wish to allow use of your version of this file only
+ * under the terms of either the GPL or the LGPL, and not to allow others to
+ * use your version of this file under the terms of the EPL, indicate your
+ * decision by deleting the provisions above and replace them with the notice
+ * and other provisions required by the GPL or the LGPL. If you do not delete
+ * the provisions above, a recipient may use your version of this file under
+ * the terms of any one of the EPL, the GPL or the LGPL.
+ ***** END LICENSE BLOCK *****/
+
+package org.jruby.ext.ractor;
+
+import org.jruby.Ruby;
+import org.jruby.ext.set.EnumerableExt;
+import org.jruby.ext.set.RubySet;
+import org.jruby.ext.set.RubySortedSet;
+import org.jruby.runtime.load.Library;
+
+public final class RactorLibrary implements Library {
+
+    public void load(Ruby runtime, boolean wrap) {
+        RactorLibrary.load(runtime);
+    }
+
+    public static void load(Ruby runtime) {
+        var context = runtime.getCurrentContext();
+        Ractor.createRactorClass(context, runtime.getObject());
+    }
+
+}
